chore: clean up frontend duplication and dead code
Extract duplicated SSE stream reading logic into a shared readSseStream()
utility, consolidate DEFAULT_FRAMEWORK into a single export from types/,
deduplicate checkpoint tree rebuild in useCheckpoint, and remove unused
Project/FileViewerProps interfaces and stepId variable. Also fixes an
off-by-one bug in SSE parsing (slice(5) → slice(6) for "data: " prefix).
